Add files via upload
This is a normal html file, so you can just double click it and open it in chrome. If you've done the previous steps (set up mamp, etc...) then this is just a simple web page with two buttons, one for researcher and one for reviewer. The researcher button redirects to the upload field, the reviewer button redirects to google randomly because I still couldn't figure out how to make the rest..